requirements.txt

# Campus Rideshare — Python Dependencies
# Install with: pip install -r requirements.txt

# Flask Core
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.1

# PostgreSQL Driver
psycopg2-binary==2.9.9

# WebSockets
websockets==12.0

# AI (OpenAI)
openai>=1.30.5

# Utilities
python-dotenv==1.0.1
python-dateutil==2.9.0
requests==2.32.3

# Dev / Testing
pytest==8.2.2